requirements.txt
numpy==1.23.5
pandas==2.0.3
scikit-learn==1.4.2
tensorflow==2.15.0
scikeras==0.12.0
scikit-optimize==0.9.0
psutil==5.9.5
mlflow==2.6.0
pyyaml==6.0
matplotlib==3.7.1
seaborn==0.12.2
#keras-tuner
joblib==1.3.2   # Explicitly added if you're using 'import joblib'

# Streamlit App
streamlit==1.18.1
altair<5

# Inference App
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.5